+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * **internal:** Support for message queue on segment creation ([CE-749](https://jira.rfcx.org/browse/CE-749))
 * **legacy:** GET v2/events endpoint uses new events database structure ([CE-868](https://jira.rfcx.org/browse/CE-868))
+* **legacy:** POST v2/events/:guid/review endpoint does nothing ([CE-904](https://jira.rfcx.org/browse/CE-904))
 
 ## 1.0.22-hotfix.0 (2021-06-16)